# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Literary Dispatch API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EntryType = (typeof EntryType)[keyof typeof EntryType];

export const EntryType = {
  diary: "diary",
  observation: "observation",
  quote: "quote",
  comment: "comment",
  literary: "literary",
} as const;

export interface Entry {
  id: number;
  title: string;
  subject: string;
  content: string;
  createdAt: string;
  type: EntryType;
}

export type CreateEntryInputType =
  (typeof CreateEntryInputType)[keyof typeof CreateEntryInputType];

export const CreateEntryInputType = {
  diary: "diary",
  observation: "observation",
  quote: "quote",
  comment: "comment",
  literary: "literary",
} as const;

export interface CreateEntryInput {
  /** @maxLength 200 */
  title: string;
  /** @maxLength 200 */
  subject: string;
  content: string;
  type: CreateEntryInputType;
  adminPassword: string;
}

export interface ArtistTheme {
  id: number;
  artistName: string;
  artistBio: string;
  era: string;
  colorPrimary: string;
  colorSecondary: string;
  colorAccent: string;
  colorBackground: string;
  colorText: string;
  fontHeading: string;
  fontBody: string;
  artworkUrls: string[];
  artworkTitles: string[];
  moodDescription: string;
  activeFrom: string;
}

export interface GenerateThemeInput {
  adminPassword: string;
  artistName?: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PasswordInput {
  password: string;
}

export interface AuthResult {
  valid: boolean;
}