# lib/api-zod/src/generated/types/artistTheme.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Literary Dispatch API
 * OpenAPI spec version: 0.1.0
 */

export interface ArtistTheme {
  id: number;
  artistName: string;
  artistBio: string;
  era: string;
  colorPrimary: string;
  colorSecondary: string;
  colorAccent: string;
  colorBackground: string;
  colorText: string;
  fontHeading: string;
  fontBody: string;
  artworkUrls: string[];
  artworkTitles: string[];
  moodDescription: string;
  activeFrom: Date;
}