# lib/api-zod/src/generated/types/createEntryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Literary Dispatch API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEntryInputType } from "./createEntryInputType";

export interface CreateEntryInput {
  /** @maxLength 200 */
  title: string;
  /** @maxLength 200 */
  subject: string;
  content: string;
  type: CreateEntryInputType;
  adminPassword: string;
}