# lib/api-zod/src/generated/types/entry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Literary Dispatch API
 * OpenAPI spec version: 0.1.0
 */
import type { EntryType } from "./entryType";

export interface Entry {
  id: number;
  title: string;
  subject: string;
  content: string;
  createdAt: Date;
  type: EntryType;
}