# lib/api-zod/src/generated/types/generateThemeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Literary Dispatch API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateThemeInput {
  adminPassword: string;
  artistName?: string;
}